requirements.txt
langchain_groq
langchain_core
python-dotenv
streamlit
langchain_community